// src/assets/model/types.ts
export type Point = {
    isStartPoint: boolean
    isStartEtap: boolean
    x: number
    y: number
}

export type Rating = {
    "id": string
    "name": string
    "lastName": string
    "img": string
    "points": string
}

export type Friend = {
    "id": string
    "name": string
    "lastName": string
    "img": string
}